weather-app-server/requirements.txt
Django==2.1.7
django-cors-headers==2.4.1
python-decouple==3.1
pytz==2018.9
